requirements.txt
python-telegram-bot>=13.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
openai>=0.27.0
langchain>=0.0.200
llama-index>=0.6.0
faiss-cpu>=1.7.0
chromadb>=0.3.0
notion-client>=1.0.0
pytest>=7.0.0
python-dotenv>=0.20.0
cryptography>=3.4.0
